exclude values which length exceeds specified limit from XML presentation FIX IllegalArgumentException of XML->DICOM converter for tags >= (8000,0000) Synchronize Dataset.calcLength and Dataset.writeDataset with ArrayList to prevent concurrent Dataset updates by putXX during write

+++ dcm4che14/trunk/src/java/org/dcm4cheri/data/SAXHandlerAdapter2.java
@@ -125,7 +125,7 @@ class SAXHandlerAdapter2 extends DefaultHandler {
     }
 
     private void onStartElement(String tagStr, String vrStr, String src) throws IOException {
-        int tag = Integer.parseInt(tagStr, 16);
+        int tag = (int) Long.parseLong(tagStr, 16);
         this.vr = vrStr == null ? VRMap.DEFAULT.lookup(tag) : VRs.valueOf(vrStr);
         handler.startElement(tag, vr, -1);
         state = EXPECT_VAL_OR_FIRST_ITEM;
